requirements.txt
# requirements.txt
opencv-python
torch
ultralytics
pdf2image
Pillow
